fixed implicit conversion warning
git-svn-id: http://svn.berlios.de/svnroot/repos/sems/trunk@529 8eb893ce-cfd4-0310-b710-fb5ebe64c474

+++ apps/click2dial/Click2Dial.cpp
@@ -105,7 +105,8 @@ AmSession* Click2DialFactory::onInvite(const AmSipRequest& req, AmArg& session_p
   string callee_uri, a_realm, a_user, a_pwd;
 
   if(session_params.size() != 4) {
-    ERROR("Need 4 parameters, got %d\n", session_params.size());
+    ERROR("Need 4 parameters, got %lu\n", 
+	  (long unsigned int)session_params.size());
     return NULL;
   }
